// services/frontend/src/domain/enums.ts
// Generated from ../api-server/docs/openapi.yaml. Do not edit.
export const apiTypes = [
  "rest",
  "graphql",
  "grpc",
  "json-rpc",
  "soap",
  "websocket",
  "odata",
  "sse",
  "event-consumer",
  "topic",
  "subject",
  "exchange",
  "queue",
  "native-protocol",
  "database",
] as const;
export const clientNames = [
  "s3-client",
  "rest-client",
  "graphql-client",
  "grpc-client",
  "grpc-streaming-client",
  "json-rpc-client",
  "soap-client",
  "websocket-client",
  "odata-client",
  "sse-client",
  "kafka-client",
  "rabbitmq-client",
  "amqp-client",
  "redpanda-client",
  "nats-client",
  "pulsar-client",
  "sqs-client",
  "gcp-pub-sub-client",
  "azure-service-bus-client",
  "redis-streams-client",
  "activemq-client",
  "ibm-mq-client",
  "native-protocol-client",
] as const;
export const languages = [
  "c",
  "c++",
  "c#",
  "go",
  "java",
  "javascript",
  "typescript",
  "python",
  "ruby",
  "php",
  "rust",
  "swift",
  "kotlin",
  "scala",
  "dart",
  "lua",
  "perl",
  "r",
  "objective-c",
  "shell",
  "bash",
  "powershell",
  "sql",
  "solidity",
  "haskell",
  "elixir",
  "erlang",
  "clojure",
  "groovy",
  "fortran",
  "cobol",
  "assembly",
  "visual-basic",
  "matlab",
  "pascal",
  "delphi",
  "f#",
  "julia",
  "zig",
  "nim",
  "ocaml",
  "prolog",
  "common-lisp",
  "scratch",
] as const;